Changed quite some stuff
Changed some in the server.py, added a battery chart of one of the agents, as well as highlighted it in blue on the grid. Also added the possibility to change the layout of the Charge poles: random, circle, latin hypercube, which comes back in model.
Added the time component, which makes cars wait at home / work / shopping. Changed the battery cost of driving, and the amount the cars charge each tilmestep. It is comparable with Tesla cars, and every tilmestep is 15 minutes. Every step in the grid is 1 km. Changed some around, to improve efficiency, and changed the way braveness is calculated. Also changed it, when cars come under certain battery level, while they don't have a target yet, they will move randomly, instead of a random point around its center. I think that was it.